disable Naming/VariableNumber check on symbols
After upgrading to latest Rubocop version (1.4.2) it started to complain:

```
[...] Naming/VariableNumber: Use normalcase for symbol numbers.
      visit = visits(:pam_cafe_20200101)
                     ^^^^^^^^^^^^^^^^^^
```

Apparently the check already existed for variable names. Now it's applied on symbols too and the default style (`normalcase`) would be this:

```
:pam_cafe20200101
```

We think this is less readable in this case, that's why we change the style to `snake_case`.